@@ -4,6 +4,8 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,7 +48,7 @@ public final class EventManager {
 
     public static final String UNLOAD = "UNLOAD";
 
-    private Map<String, Set<EventListener>> eventListeners = new HashMap<String, Set<EventListener>>();
+    private Map<String, Set<EventListener>> eventListeners = new ConcurrentHashMap<>();
 
     /**
      *
@@ -57,7 +59,7 @@ public final class EventManager {
         for (String currentEvent : events) {
             Set<EventListener> listeners = eventListeners.get(currentEvent);
             if (listeners == null) {
-                listeners = new HashSet<>();
+                listeners = ConcurrentHashMap.newKeySet();
                 eventListeners.put(currentEvent, listeners);
             }
             listeners.add(eventListener);
